Changed loop begin/end/increment to integers + made increment optional with default value set to 1

+++ source/include/dqm4hep/XMLParser.h
@@ -178,7 +178,7 @@ namespace dqm4hep {
       
       // for loop related
       void resolveForLoops(TiXmlElement *element);
-      void resolveForLoop(TiXmlNode *node, const std::string &id, float value);
+      void resolveForLoop(TiXmlNode *node, const std::string &id, int value);
 
     private:
       typedef std::map<std::string, std::shared_ptr<DBInterface>> DBInterfaceMap;
